wip, splitting out vault cipher impls
